reorganizing all the tests
